Updated to 1.7, lots of changes, see rest of log for details.
-Rewrote the Recompiler Reg Caching, now uses 7 "live" registers. Needed pretty much a complete rewrite of all ops.
-Fixed 1.3 spec Recompiler operations under DIV
-Fixed ini saving when the emulator crashes
-Updated the project to VS 2015
-Fixed various compiler warnings
-Added Bruised game by me, a really bad fighting game.
-Replaced the ADSR test with a working version
-Added Flip Offscreen test rom for testing your graphics system.
-Added Spec 1.3 tests. You still need BC_TestRom for the normal ones.

There is room for optimisation in the Recompiler changes, I could reduce the flushing of the registers a fair bit, but figured I'd commit this first and work on that later.

+++ RefChip16/D3DApp.h
@@ -25,8 +25,8 @@
 #include <SDL.h>
 #include <SDL_syswm.h>
 
-extern int SCREEN_WIDTH;
-extern int SCREEN_HEIGHT;
+extern unsigned short SCREEN_WIDTH;
+extern unsigned short SCREEN_HEIGHT;
 
 extern char MenuVSync;
 extern char Smoothing;
